requirements.txt
asgiref==3.8.1
astroid==3.2.2
autopep8==2.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
dill==0.3.8
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
packaging==24.1
pep8==1.7.1
platformdirs==4.2.2
pycodestyle==2.12.0
pylint==3.2.5
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tomlkit==0.12.5
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
